Use file-dropzone component for cover image upload
Switch the manager event-settings page and the proposer wizard final-step
to render `cover_image` via tessera_field, which uses the shared dropzone
component from #227. The custom preview script and ad-hoc <img> markup
are dropped.

Errors are now rendered INSIDE the dropzone (in the empty-state slot),
with a red border and a triangle icon. Tessera skips its external
errors render for FileInput so the message appears once, in place.

The dropzone init JS is moved to its own bundle (src/dropzone.ts) so it
can be loaded on any page that uses the component, not just the
encounter form.

render_file_input now accepts a string URL as initial value, so the
existing cover (which the view holds as a string DTO field, not a
FieldFile) renders as the dropzone's initial preview.
