+++ src/ludamus/adapters/web/django/templatetags/tessera/file_input.py
@@ -2,13 +2,25 @@
 
 from __future__ import annotations
 
+from posixpath import basename
 from typing import TYPE_CHECKING
+from urllib.parse import unquote, urlsplit
 
 from django.forms import ImageField
 from django.template.loader import render_to_string
 
 if TYPE_CHECKING:
     from django.forms import BoundField
+
+
+def _initial_url_and_name(initial: object) -> tuple[str | None, str]:
+    if not initial:
+        return None, ""
+    if (url := getattr(initial, "url", None)) is not None:
+        return url, str(initial)
+    if isinstance(initial, str):
+        return initial, unquote(basename(urlsplit(initial).path))
+    return None, ""
 
 
 def render_file_input(field: BoundField) -> str:
@@ -21,8 +33,7 @@ def render_file_input(field: BoundField) -> str:
     accept = attrs.get("accept") or (
         "image/*" if isinstance(field.field, ImageField) else ""
     )
-    initial = field.value()
-    initial_url = getattr(initial, "url", None) if initial else None
+    initial_url, initial_name = _initial_url_and_name(field.value())
     return render_to_string(
         "components/file-dropzone.html",
         {
@@ -31,8 +42,9 @@ def render_file_input(field: BoundField) -> str:
             "required": field.field.required,
             "accept": accept,
             "is_image": isinstance(field.field, ImageField),
+            "errors": field.errors,
             "has_errors": bool(field.errors),
             "initial_url": initial_url,
-            "initial_name": str(initial) if initial_url else "",
+            "initial_name": initial_name,
         },
     )
