+++ src/ludamus/templates/chronology/propose/parts/review.html
@@ -1,6 +1,6 @@
 {% load i18n %}
 {% load cfp_tags %}
-{% load tessera %}
+{% load tessera django_vite %}
 <div id="wizard-content">
     {% include "chronology/propose/parts/_stepper.html" %}
     {% if review|default:"" %}
@@ -166,28 +166,7 @@
               hx-target="this"
               hx-swap="none">
             {% csrf_token %}
-            <div class="pt-6">
-                <label for="id_cover_image"
-                       class="block text-sm font-medium text-foreground-secondary mb-2">
-                    {% translate "Cover image" %}
-                </label>
-                <img id="cover-image-preview"
-                     alt=""
-                     width="960"
-                     height="540"
-                     class="mb-3 aspect-video w-full rounded-xl object-cover border border-border"
-                     hidden>
-                <input type="file"
-                       name="cover_image"
-                       id="id_cover_image"
-                       accept="image/jpeg,image/png,image/webp,image/avif"
-                       data-preview-target="cover-image-preview"
-                       class="w-full px-4 py-2 text-sm rounded-lg border border-border bg-bg-secondary text-foreground focus:outline-none focus:ring-2 focus:ring-primary">
-                <p class="text-sm text-foreground-muted mt-1">{{ image_form.cover_image.field.help_text }}</p>
-                {% if image_form.cover_image.errors %}
-                    <p class="text-sm text-danger mt-1">{{ image_form.cover_image.errors.0 }}</p>
-                {% endif %}
-            </div>
+            <div class="pt-6">{% tessera_field image_form.cover_image %}</div>
             <div class="flex flex-col-reverse sm:grid sm:grid-cols-2 gap-3 pt-6">
                 <button type="button"
                         class="btn btn-secondary py-3"
@@ -198,6 +177,6 @@
                 <button type="submit" class="btn btn-primary py-3">{% translate "Submit Proposal" %}</button>
             </div>
         </form>
-        {% include "components/cover_image_preview_script.html" %}
+        {% vite_asset 'src/dropzone.ts' %}
     {% endif %}
 </div>
