+++ src/ludamus/templates/panel/settings.html
@@ -1,6 +1,6 @@
 {% extends "panel/base.html" %}
 {% load i18n %}
-{% load tessera %}
+{% load tessera django_vite %}
 {% block title_prefix %}
     {% translate "Event Settings" %} -
 {% endblock title_prefix %}
@@ -76,24 +76,7 @@
                                           rows="3"
                                           class="w-full border border-border rounded-lg px-4 py-2 focus:outline-none focus:ring-2 focus:ring-primary">{{ form.description.value|default:'' }}</textarea>
                             </div>
-                            <div>
-                                <label class="block text-sm font-medium text-foreground-secondary mb-2"
-                                       for="id_cover_image">{% translate "Cover image" %}</label>
-                                <img id="cover-image-preview"
-                                     src="{{ current_event.cover_image_url|default:'' }}"
-                                     alt=""
-                                     width="960"
-                                     height="540"
-                                     class="mb-3 aspect-video w-full rounded-xl object-cover border border-border"
-                                     {% if not current_event.cover_image_url %}hidden{% endif %}>
-                                <input type="file"
-                                       name="cover_image"
-                                       id="id_cover_image"
-                                       accept="image/jpeg,image/png,image/webp,image/avif"
-                                       data-preview-target="cover-image-preview"
-                                       class="w-full border border-border rounded-lg px-4 py-2 focus:outline-none focus:ring-2 focus:ring-primary">
-                                <p class="mt-1 text-sm text-foreground-muted">{{ form.cover_image.field.help_text }}</p>
-                            </div>
+                            {% tessera_field form.cover_image %}
                         </div>
                     </div>
                 </div>
@@ -136,5 +119,5 @@
 {% endblock content %}
 {% block extra_scripts %}
     {{ block.super }}
-    {% include "components/cover_image_preview_script.html" %}
+    {% vite_asset 'src/dropzone.ts' %}
 {% endblock extra_scripts %}
